+++ packages/frontend/src/components/tvl-breakdown/TvlBreakdownSummaryBox.stories.tsx
@@ -1,7 +1,8 @@
 import { Story } from '@storybook/react'
 import React from 'react'
-import { TvlBreakdownSummaryBox as TvlBreakdownSummaryBoxComponent } from './TvlBreakdownSummaryBox'
+
 import { TvlBreakdownViewProps } from '../../pages/scaling-projects-tvl-breakdown/view/TvlBreakdownView'
+import { TvlBreakdownSummaryBox as TvlBreakdownSummaryBoxComponent } from './TvlBreakdownSummaryBox'
 
 export default {
   title: 'Components/TVL Breakdown/TvlBreakdownSummaryBox',
