+++ packages/frontend/src/pages/scaling-projects-tvl-breakdown/props/getTvlBreakdownHeader.ts
@@ -1,16 +1,13 @@
 import { Layer2 } from '@l2beat/config'
+import { ProjectAssetsBreakdownApiResponse } from '@l2beat/shared-pure'
+
 import { TvlBreakdownHeaderProps } from '../view/TvlBreakdownHeader'
-import {
-  DetailedTvlApiResponse,
-  ProjectAssetsBreakdownApiResponse,
-  TvlApiResponse,
-} from '@l2beat/shared-pure'
 
 export function getTvlBreakdownHeader(
   project: Layer2,
-  tvlBreakdownApiResponse: ProjectAssetsBreakdownApiResponse | undefined,
+  tvlBreakdownApiResponse: ProjectAssetsBreakdownApiResponse,
 ): TvlBreakdownHeaderProps {
-  const date = tvlBreakdownApiResponse?.dataTimestamp.toDate() || new Date()
+  const date = tvlBreakdownApiResponse.dataTimestamp.toDate()
   const month = date.getMonth() + 1
   const tvlBreakdownDate = `${date.getFullYear()}-${
     month.toString().length === 1 ? `0${month}` : month
